Add Images in framework-res.apk
Just to get things moving added in images and hex colors used in the Triumph Rom ware translatable. For the most part the same images used in the original Rom are also found in the “framework-res.apk” but rarely used in the UI. Overall a few images are used from the “framework-res.apk” but the majority of the images used in the overall UI seem to come from various places like for example “com.htc.resources.apk” but is a start.